Create helperFuncs.js
