+++ src/redux/store.js
@@ -3,7 +3,6 @@ import allCoinsReducer from './all-coins/allCoinsSlice';
 import coinDetailsReducer from './coin-details/coinDetailsSlice';
 import categoriesReducer from './categories/categoriesSlice';
 import globalDataReducer from './global-data/globalDataSlice';
-import PriceConvertionReducer from './price-convertion/priceConvertionSlice';
 
 const store = configureStore({
   reducer: {
@@ -11,7 +10,6 @@ const store = configureStore({
     coinDetails: coinDetailsReducer,
     categories: categoriesReducer,
     globalData: globalDataReducer,
-    priceConvertion: PriceConvertionReducer,
   },
 });
